+++ libsrc/ffdec_lib/src/com/jpexs/decompiler/flash/action/ActionLocalData.java
@@ -37,6 +37,8 @@ public class ActionLocalData extends BaseLocalData {
 
     public boolean insideDoInitAction;
 
+    public boolean insideDefineFunction1 = false;
+
     public ActionLocalData(boolean insideDoInitAction) {
         regNames = new HashMap<>();
         variables = new HashMap<>();