+++ libsrc/ffdec_lib/src/com/jpexs/decompiler/flash/action/swf3/ActionPlay.java
@@ -52,7 +52,7 @@ public class ActionPlay extends Action {
     }
 
     @Override
-    public void translate(boolean insideDoInitAction, GraphSourceItem lineStartAction, TranslateStack stack, List<GraphTargetItem> output, HashMap<Integer, String> regNames, HashMap<String, GraphTargetItem> variables, HashMap<String, GraphTargetItem> functions, int staticOperation, String path) {
+    public void translate(boolean insideDoInitAction, boolean insideDefineFunction1, GraphSourceItem lineStartAction, TranslateStack stack, List<GraphTargetItem> output, HashMap<Integer, String> regNames, HashMap<String, GraphTargetItem> variables, HashMap<String, GraphTargetItem> functions, int staticOperation, String path) {
         if (!output.isEmpty() && (output.get(output.size() - 1) instanceof GotoFrameActionItem)) {
             GotoFrameActionItem gta = (GotoFrameActionItem) output.remove(output.size() - 1);
             output.add(new GotoFrame2ActionItem(this, lineStartAction, new DirectValueActionItem(gta.frame + 1), false, true, 0));